+++ src/utils/dateUtils.ts
@@ -19,6 +19,9 @@ import dayjs from 'dayjs';
 import localizedFormat from 'dayjs/plugin/localizedFormat';
 import localeData from 'dayjs/plugin/localeData';
 import utc from 'dayjs/plugin/utc';
+import createDebug from 'debug';
+
+const debug = createDebug('enchanted-web-components:utils:dateUtils.ts');
 
 dayjs.extend(localizedFormat);
 dayjs.extend(localeData);
@@ -128,9 +131,7 @@ export const getLocalizedDays = (locale: string = DEFAULT_CALENDAR_LOCALE): stri
       });
     } else if (locale === 'ar') {
       localizedDays = localizedDays.map((day, index) => {
-        // eslint-why - Enchanted-logger could not be used, later we should switch to the debug package
-        // eslint-disable-next-line no-console
-        console.debug(`Mapping day string: ${day}, with index: ${index}`);
+        debug('Mapping day string: %s, with index: %d', day, index);
         if (index === 0) {
           return day.slice(4, 5); // For Monday: Align with Enchanted React DatePicker implementation
         }
@@ -462,12 +463,8 @@ export const correctDateEndForDatePickerFilter = (dateEnd: string): string => {
       newDateEnd = getUnixTimestampMilliseconds(`${String(Number(year) + 1).padStart(4, '0')}-01-01`).toString();
     }
 
-    // eslint-why - Enchanted-logger could not be used, later we should switch to the debug package
-    // eslint-disable-next-line no-console
-    console.log(`The dateEnd passed to the util is: ${formatDate(Number(dateEnd), FORMAT_FOR_CONVERTING_TO_UNIX_TIMESTAMP)}`);
-    // eslint-why - Enchanted-logger could not be used, later we should switch to the debug package
-    // eslint-disable-next-line no-console
-    console.log(`The newDateEnd calculated is: ${formatDate(Number(newDateEnd), FORMAT_FOR_CONVERTING_TO_UNIX_TIMESTAMP)}`);
+    debug('The dateEnd passed to the util is: %s', formatDate(Number(dateEnd), FORMAT_FOR_CONVERTING_TO_UNIX_TIMESTAMP));
+    debug('The newDateEnd calculated is: %s', formatDate(Number(newDateEnd), FORMAT_FOR_CONVERTING_TO_UNIX_TIMESTAMP));
   }
 
   return newDateEnd;